Test util refactor
Pure refactoring of the test classes - moving some method to the parent
class so the subsequent commits can use them. Extracting methods to a
new class so it can be consumed by later changes.
